python/pdf-helpers/requirements.txt
tk==0.1.0
PyPDF2==3.0.1
tabula-py==2.9.0
tabulate==0.9.0
